Fix broken Windows installation instructions (#29574)
I'm experiencing problems with the "Install the fleetctl command line
tool" step on https://fleetdm.com/try-fleet for trying out Fleet hosting
on Windows. The root cause seem to be a mismatch between the Windows
ZIP-file naming in the script vs. on
https://api.github.com/repos/fleetdm/fleet/releases/latest

I was able to overcome the problem by changing `_windows.zip` to
`_windows_amd64.zip` in the script.