Product design rituals: Design sprint review/kickoff -> Product design sprint review/kickoff (#29987)
Why? Because @mike-j-thomas is the DRI for the visual/brand design and
it makes that distinction a bit clearer. (Basically, anything that’s
heavier on the visual design and not product design, e.g. design system
stuff, should involve him, and sounds like we should try to loop him in
more going forward to make sure there’s not a mismatch between the
product itself and the brand.)