Fix issue saving labels (#30126)
# Details 
This PR fixes an issue where the user is redirected to a 500 error page
after saving a new label if they were on the All Teams host page before
creating the label. The issue was due to the redirect having
`team_id=-1` in it, and the fix is to detect that All Teams is selected
and, if so, not set `team_id` in the URL at all.

# Checklist
<!-- Note that API documentation changes are now addressed by the
product design team. -->
- [X] Manual QA for all new/changed functionality
- [X] For unreleased bug fixes in a release candidate, confirmed that
the fix is not expected to adversely impact load test results or alerted
the release DRI if additional load testing is needed.

# Testing

Tested adding a new and dynamic label from All Teams, No team and a
custom team. Vids below for "All Team":

Before:

![30036-before](https://github.com/user-attachments/assets/f958f31d-9b8d-4a61-8b3d-8d55590628fe)

After:

![30036-after](https://github.com/user-attachments/assets/0b424ed7-9bab-45cd-b4e5-2a6ae4a1309a)