Cherrypick: Fix bug for not being to reenable end user migration in UI (#30106) (#30191)
Relates to #30063

Cherry pick PR: This fixes an issue added in the
[PR](https://github.com/fleetdm/fleet/pull/29968) where the user was not
able to reenable the end user migration form.

I've also added improved a11y attributes to the slider component,
ensured we are functionally disabling the form controls during gitops
mode and not just visually, and updated/added tests for the
EndUserMigrationSection component.