Use dedicated, string-interpolated queries for single-host MDM status checks to reduce prepared statement usage
The hottest path for these changes is the Orbit config getter, which runs every 30 seconds for each host. That means that for 10k enrolled hosts this will save ~333 prepares per second...which adds up.

There are a few other places that use this query, but not on as hot of a path