pkg/open: carryover some more env vars to launched xdg-open process
Fleet Desktop fails to launch a browser in Ubuntu 24.04 LTS when the
browser is installed as a snap, due to `.desktop` files installed by
snapd being in a location which is only discoverable through the
`XDG_DATA_DIRS` environment variable.

To workaround this we can grab some additional environment variables
from the Xwayland process to enable discovery of the `.desktop` files
and use of the user's preferences in the currently running DE.

The list of variables to grab was found through trial-and-error and
seems to be the minimum required for `xdg-open` to function.

Refactored and optimized the code to fetch environment variables from a
running process to reuse search results, match only processes running as
the current user, and robustly handle multiple process results.

Fixes #19043.