pkg/open: fix launching browser in some scenarios
Fleet Desktop may fail to launch a browser in Ubuntu 24.04 LTS under
some scenarios, including Wayland sessions and/or when the browser is
installed as a snap or flatpak.

The previous approach of searching for an Xwayland process from which to
extract environment variables is incomplete. When the browser is
installed as a snap, `.desktop` files are installed by snapd into a
location which is only discoverable through the `XDG_DATA_DIRS`
environment variable.

To workaround this we need to change the approach in several ways:

- Look for `xdg-desktop-portal` processes, not just `Xwayland`. Both
  Firefox and Chrome support native operation under Wayland now, so it's
  not correct to rely upon the existence of an Xwayland process alone.
- Copy several more environment varibles: `XDG_CURRENT_DESKTOP`,
  `XDG_RUNTIME_DIR`, `XDG_DATA_DIRS`, and `PATH`, in addition to
  `XAUTHORITY`.

These variables allow fleet-desktop to honor the user's browser
preference according to the currently running DE.

The list of variables to grab was found through trial-and-error and
seems to be the minimum required for `xdg-open` to function.

Refactored and optimized the code to fetch environment variables from a
running process to match only processes running as the current user and
robustly handle multiple process results.

Fixes #19043, #27209.