[gitops] support multiple SoftwarePackageSpecs in one yaml file
The list of software configured via GitOps might be dynamic, for example if the GitOps pipeline builds and pushes packages just before fleet gitops is called, or if partial package lists are shared between teams.

In these cases, it's useful to be able to include one file containing multiple packages. This PR refactors `parseSoftware` to support that use case.

This remains backward compatible with the old style of using one package per included file. This is verified by a new test, `TestSoftwarePackagesUnmarshalMulti`.

Signed-Off-By: Dan Fuhry <dan@fuhry.com>