Add note to macOS MDM setup about failures during Setup Assistant (#30897)
For #30089 . This behavior appears to be an Apple bug, or at least an
unsupported usecase(rebooting while at the Remote Management screen,
before the DeviceConfigured is sent) and thus our best course of action
is to document that a user should Wipe/Reset the device when this
happens unless we learn more about how to prevent this which seems
unlikely.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Documentation**
* Added a new section detailing potential causes and recovery steps for
automatic enrollment failures during macOS MDM setup.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Marko Lisica <83164494+marko-lisica@users.noreply.github.com>
Co-authored-by: Rachael Shaw <r@rachael.wtf>