[ee/server] support delivering certificates through scripts
In the absence of a fully fleshed out certificate delivery feature for
Linux hosts (#29577), another option is to simply perform variable
substitution when hosts download scripts to execute them.

This PR adds support for variable substitution, including SCEP and
DigiCert certificate variables, into scripts.

No documentation or tests are supplied as this is intended as a stop-gap
for our internal use, but I'm putting it up as a benefit for the
community. Variable syntax is exactly the same as that which is
described on the [existing
guide](https://fleetdm.com/guides/connect-end-user-to-wifi-with-certificate).

Example scripts:

- Zscaler + DigiCert:
  https://gist.github.com/fuhry/554b105015fd8a03d84ce8351a80a040
- SCEP (uses [micromdm scep client](https://github.com/micromdm/scep)):
  https://gist.github.com/fuhry/454c117f9cbb7c97cdb5138d6e9cc693

Significant code is copied from `apple_mdm.go` with apple-specific bits,
such as profile UUID, stripped out. This obviously has significant tech
debt/maintainability implications if merged as-is. Unfortunately I lack
the free time to refactor this in the way that is truly needed, which is
a common fleetvar->value mapping and all the apple XML- and
shell-script-specific bits isolated.

In addition, the apple version of this variable substitution was made to
scale, by pre-populating variables in the database in order to optimize
delivery at the time the profile is fetched. This PR interpolates at the
time the script is fetched by orbit to be executed, potentially
increasing latency of the script download endpoint.

As such, I have no expectation that this will be merged as-is, but have
contributed it as a possible starting point for future feature work that
is already on the roadmap.

Signed-Off-By: Dan Fuhry <dan@fuhry.com>