add basic handling for vpp app installs on osquery perf (#31178)
# Checklist for submitter

Adds some very basic VPP app install functionality to osquery-perf.

When one of the supported apps (in this rev: Evernote, Bear, Craft,
Goodnotes) is installed, then OSQP

- Handles the `InstallApplication` command and responds with an Ack
- Handles the first verification `InstalledApplicationList` command and
marks the app as "installed" internally, but does not return the app in
the response
- Handles the second verification `InstalledApplicationList` and returns
the app in the response. This should verify it as installed on the Fleet
side.

If some of the following don't apply, delete the relevant line.

<!-- Note that API documentation changes are now addressed by the
product design team. -->

- [x] Manual QA for all new/changed functionality