Pinning to old MS office uninstall script (#31233)
homebrew changed the formula for microsoft word, in particular the zap
section, to delete more files than we think is correct.
https://github.com/Homebrew/homebrew-cask/pull/218350

This uninstall script is what we had prior to
7a1e469ac09dd28944e9cbfeb55088229a3ff488 Microsoft office FMA version
from release notes (#30686), which used the updated homebrew formula.

Related to #30082

# Checklist for submitter

- [x] Manual QA for all new/changed functionality
- [x] For unreleased bug fixes in a release candidate, confirmed that
the fix is not expected to adversely impact load test results or alerted
the release DRI if additional load testing is needed.