UI: `InventoryVersions` component (#31253)
## This is a foundational UI component shared between my current dev
work and @RachelElysia's. This PR is to share that component while we
are both developing on top of it.


Much of this logic exists in the current `SoftwareDetailsModal` (under
the hosts details page, not the identically named Activity feed
component - will be clarifying that naming in an upcoming PR). This PR
is the first step in a reorganization of much of these abstractions to
better suit the desired UX around software modals. Much more detail to
come shortly in actual feature work in upcoming PRs from both Rachel and
I.

---------

Co-authored-by: Jacob Shandling <jacob@fleetdm.com>