Add edge case scenario and how to resolve it. (#31655)
### Summary  
This PR adds documentation for an edge case in Intune: certain policies
may appear in the NodeCache without a proper Boolean value
(`ExpectedValue: -1`), which can lead to confusion when verifying the
CSP policy application.

### What changed  
- Added a **note** explaining the `-1` value in the NodeCache.  
- A PowerShell example was included to detect the issue.  
- Added guidance on verifying the runtime state of the setting.  
- Linked to the related GitHub
[issue](https://github.com/fleetdm/confidential/issues/9817) for further
context.

### Why this is needed  
When dogfooding the original guide, we identified that some policies
applied via Intune don’t explicitly set a value in the registry. Without
this clarification, users may get failures when creating CSPs for Fleet
as it expects a positive value. This update helps avoid confusion and
provides a reliable way to verify policy state.