Update run script modal (#31604)
for #30526 

# Details

This PR updates the "Run batch script" modal to allow for scheduling
scripts in the future. When a script is selected from the list, rather
than running immediately, it now presents a choice of "Run now" and
"Schedule for later". If the user opts to schedule for later, they are
presented with a form for choosing the date and time to run the script.

Run now example:


https://github.com/user-attachments/assets/6fe3a77e-a271-4b58-abfd-c9ad9196d0b5

Schedule for later example:



https://github.com/user-attachments/assets/ca43eab6-03f6-4c0b-929c-aabfb69f2e5f




# Checklist for submitter

## Testing

- [X] Added/updated automated tests
- [X] QA'd all new/changed functionality manually